improved certmanager deploy code
The bootstrap apiserver start/stop functionality has been extracted into
a single script that is also easy to use manually.
The logic for deciding whether to use this script or not has been moved
into the certmanager deploy script. The script will only be called if
one master node is detected. The detection is now done via kubectl
instead of terraform.